Make PID file creating errors fatal
Adaptation of btc@561e375c73a37934fe77a519762d81edf7a3325c and btc@3782075a5fd4ad0c15a6119e8cdaf136898f679e
